+++ lua/LUA/ak/demo-anlagen/tutorial-ampel/Andreas_Kreuz-Tutorial-Ampelkreuzung-2-main.lua
@@ -146,7 +146,7 @@ k1:setSwitchInStrictOrder(true)
 local ModuleRegistry = require("ak.core.ModuleRegistry")
 ModuleRegistry.registerModules(
     require("ak.core.CoreLuaModule"),
-    require("ak.road.CrossingLuaModul")
+    require("ak.road.CrossingLuaModule")
 )
 
 function EEPMain()